core: Remove default aspect ratio limit for old apps
I have never encountered a single app that this aspect ratio limit is
actually helpful for. Most legacy apps are able to cope with longer
aspect ratios fine — perhaps not optimally, but it looks better than
having a big black box at the bottom of the screen.

Instead of limiting older apps' screen space unconditionally or exposing
it in Settings, remove the limit entirely so the user doesn't have to go
through the trouble to begin with. Note that apps explicitly declaring a
shorter aspect ratio will still get their constraint satisfied.

Change-Id: Ic415288b24129a634476b5cc01a6f202588f4b89
Signed-off-by: rk134 <rahul-k@bigdi.cc>